Adding args option to RunTestExecutable task for the native test suite
