+++ subprojects/docs/src/docs/dsl/org.gradle.api.tasks.Exec.xml
@@ -7,36 +7,6 @@
                     <td>Name</td>
                 </tr>
             </thead>
-            <tr>
-                <td>args</td>
-            </tr>
-            <tr>
-                <td>commandLine</td>
-            </tr>
-            <tr>
-                <td>executable</td>
-            </tr>
-            <tr>
-                <td>workingDir</td>
-            </tr>
-            <tr>
-                <td>environment</td>
-            </tr>
-            <tr>
-                <td>standardInput</td>
-            </tr>
-            <tr>
-                <td>standardOutput</td>
-            </tr>
-            <tr>
-                <td>errorOutput</td>
-            </tr>
-            <tr>
-                <td>ignoreExitValue</td>
-            </tr>
-            <tr>
-                <td>execResult</td>
-            </tr>
         </table>
     </section>
     <section>
@@ -47,21 +17,6 @@
                     <td>Name</td>
                 </tr>
             </thead>
-            <tr>
-                <td>commandLine</td>
-            </tr>
-            <tr>
-                <td>args</td>
-            </tr>
-            <tr>
-                <td>executable</td>
-            </tr>
-            <tr>
-                <td>workingDir</td>
-            </tr>
-            <tr>
-                <td>environment</td>
-            </tr>
         </table>
     </section>
 </section>
